cleaned directory added sql dir
