+++ lib/models.rb
@@ -9,8 +9,9 @@ ActiveRecord::Base.logger = Logger.new(STDOUT)
 
 class Album < ActiveRecord::Base
   belongs_to :label
-  has_many :albums_tracks
-  has_many :tracks, through: :albums_tracks
+  belongs_to :genre
+  has_many :album_tracks
+  has_many :tracks, through: :album_tracks
 
   def to_s
     title + ' ' + tracks.map { |track| track.to_s}.join('/n')
@@ -21,25 +22,23 @@ class Label < ActiveRecord::Base
   has_many :albums
 end
 
-class AlbumsTrack < ActiveRecord::Base
+class AlbumTrack < ActiveRecord::Base
   belongs_to :album
   belongs_to :track
 end
 
 class Artist < ActiveRecord::Base
-
   has_and_belongs_to_many :tracks
   validates :name, presence: true, uniqueness: true
 end
 
 class Genre < ActiveRecord::Base
-  has_many :tracks
-  belongs_to :track
+  has_many :albums
+  belongs_to :album
 end
 
 class Track < ActiveRecord::Base
   has_and_belongs_to_many :artists
-  belongs_to :genre
   def to_s
     myArtists = artists.map(&:name).join(',')
     "#{myArtists} - #{title}"
