--- sqlstuff.sql
@@ -1,106 +1,0 @@
-INSERT INTO artists VALUES
-(0,'The Strokes','UK'),
-(1,'Love','UK'),
-(2,'The Velvet Underground','UK'),
-(3,'Elliott Smith','UK'),
-(4,'The Strokes','UK'),
-(5,'Jefferson Airplane','UK'),
-(6,'Jeff Buckley','UK'),
-(7,'The Strokes','UK'),
-(8,'The Velvet Underground','UK'),
-(9,'Elliott Smith','UK'),
-(10,'Elliott Smith','UK'),
-(11,'Elliott Smith','UK'),
-(12,'The Velvet Underground','UK'),
-(13,'Radiohead','UK'),
-(14,'Radiohead','UK'),
-(15,'Aphex Twin','UK'),
-(16,'Aphex Twin','UK'),
-(17,'Tame Impala','UK'),
-(18,'The Rolling Stones','UK'),
-(19,'Pixies','UK'),
-(20,'Pixies','UK'),
-(21,'Radiohead','UK'),
-(22,'Elliott Smith','UK'),
-(23,'Bob Dylan','UK'),
-(24,'Arctic Monkeys','UK'),
-(25,'Bob Dylan','UK'),
-(26,'Link Wray','UK'),
-(27,'Johnny Cash','UK'),
-(28,'James Brown','UK'),
-(29,'Nirvana','UK'),
-(30,'Aphex Twin','UK'),
-(31,'Chris Isaak','UK'),
-(32,'Led Zeppelin','UK'),
-(33,'Warpaint','UK'),
-(34,'Ms. Lauryn Hill','UK'),
-(35,'Ty Segall','UK'),
-(36,'The Clash','UK'),
-(37,'Beach Fossils','UK'),
-(38,'The Brian Jonestown Massacre','UK'),
-(39,'The Brian Jonestown Massacre','UK'),
-(40,'Erik Satie','UK'),
-(41,'The Sonics','UK'),
-(42,'Leonard Cohen','UK'),
-(43,'Wendy Rene','UK'),
-(44,'Bob Dylan','UK'),
-(45,'Deep Puddle Dynamics','UK'),
-(46,'Disciples','UK'),
-(47,'Kiasmos','UK'),
-(48,'Dominik Eulberg','UK'),
-(49,'Alias','UK'),
-(50,'Markus Acher','UK'),
-(51,'Brian Protheroe','UK'),
-(52,'The Rolling Stones','UK'),
-(53,'Wu-Tang Clan','UK'),
-(54,'King Geedorah','UK'),
-(55,'Mr Fantastik','UK'),
-(56,'Jonny Greenwood','UK'),
-(57,'Elliott Smith','UK'),
-(58,'Talking Heads','UK'),
-(59,'Francis Lai','UK'),
-(60,'Elliott Smith','UK'),
-(61,'The Raveonettes','UK'),
-(62,'The Smiths','UK'),
-(63,'Beach Fossils','UK'),
-(64,'Elliott Smith','UK'),
-(65,'Warpaint','UK'),
-(66,'J Dilla','UK'),
-(67,'Roots Manuva','UK'),
-(68,'Ian Brown','UK'),
-(69,'Wu-Tang Clan','UK'),
-(70,'Method Man','UK'),
-(71,'The Notorious B.I.G.','UK'),
-(72,'Kelis','UK'),
-(73,'T. Rex','UK'),
-(74,'R.E.M.','UK'),
-(75,'The Libertines','UK'),
-(76,'Van Morrison','UK'),
-(77,'Radiohead','UK'),
-(78,'Arctic Monkeys','UK'),
-(79,'The Libertines','UK'),
-(80,'Joy Division','UK'),
-(81,'The Smiths','UK'),
-(82,'The Smiths','UK'),
-(83,'Roots Manuva','UK'),
-(84,'Roots Manuva','UK'),
-(85,'Riddla','UK'),
-(86,'Wu-Tang Clan','UK'),
-(87,'The Stooges','UK'),
-(88,'Iggy Pop','UK'),
-(89,'The Stooges','UK'),
-(90,'Jamiroquai','UK'),
-(91,'Metallica','UK'),
-(92,'Sonic Youth','UK'),
-(93,'Jefferson Airplane','UK'),
-(94,'Calyx','UK'),
-(95,'Led Zeppelin','UK'),
-(96,'Ian Brown','UK'),
-(97,'Radiohead','UK'),
-(98,'Massive Attack','UK'),
-(99,'The Jesus and Mary Chain','UK'),
-(100,'The Velvet Underground','UK'),
-(101,'Nico','UK'),
-(102,'Television','UK'),
-(103,'The Stooges','UK'),
-(104,'Joy Division','UK');
