--- tests.rb
@@ -1,5 +1,0 @@
-require './models'
-
-puts Album.all.each do |a|
-  puts a.name
-end
